+++ PluginManager/Implementations/ZipPluginPackageResolver.cs
@@ -37,7 +37,7 @@ public sealed class ZipPluginPackageResolver<TPluginManifest> : IPluginPackageRe
 			yield break;
 		}
 
-		ZipArchive archive = new(this.ZipFile.OpenRead(), ZipArchiveMode.Read, leaveOpen: true);
+		var archive = new ZipArchive(this.ZipFile.OpenRead(), ZipArchiveMode.Read, leaveOpen: true);
 		foreach (var resolveResult in this.ResolverFactory(ZipDirectoryInfo.From(archive)).ResolvePluginPackages())
 			yield return resolveResult.Match<PluginPackageResolveResult<TPluginManifest>>(
 				success => new PluginPackageResolveResult<TPluginManifest>.Success