+++ PluginManager/Interfaces/Files/IDirectoryInfoWrapper.cs
@@ -1,0 +1,13 @@
+namespace Nanoray.PluginManager;
+
+/// <summary>
+/// Describes a wrapper <see cref="IDirectoryInfo"/>.
+/// </summary>
+public interface IDirectoryInfoWrapper : IDirectoryInfo, IFileSystemInfoWrapper
+{
+	/// <inheritdoc cref="IFileSystemInfoWrapper.Wrapped"/>
+	new IDirectoryInfo Wrapped { get; }
+
+	IFileSystemInfo IFileSystemInfoWrapper.Wrapped
+		=> this.Wrapped;
+}